1.23 성별 방문객 테이블에서 연도별 합산 상위 7개 나라 조회 완성.

+++ src/main/java/github/tourism/web/dto/statistic/GenderResponseDTO.java
@@ -17,8 +17,8 @@ public class GenderResponseDTO {
     private Integer totalPopulation;
     private Integer previousTotalPopulation;
     private BigDecimal totalIncrement;
-    private BigDecimal malePopulation;
+    private Integer malePopulation;
     private BigDecimal maleRatio;
-    private BigDecimal femalePopulation;
+    private Integer femalePopulation;
     private BigDecimal femaleRatio;
 }
